combined different scripts
